LayoutTests/fast/viewport/viewport-70-expected.txt

CONSOLE MESSAGE: line 3: Viewport width or height set to physical device width, try using "device-width" constant instead for future compatibility.
viewport size 100x110 scale 3.200000 with limits [3.200000, 5.000000]
